move api to views folder
